backend/requirements.txt
fastapi[standard]
uvicorn
pydantic
python-dotenv
openai
supabase
python-multipart
httpx
